consumer/requirements.txt
pandas==2.3.0
kafka-python==2.2.13
pydantic==2.11.7
numpy==2.3.0
psycopg2-binary==2.9.10